cmp tests refactor
